Big reorganization, error dumping implemented.
