+++ viewser/commands/__init__.py
@@ -1,0 +1,8 @@
+
+from .tables import tables
+from .queryset import queryset
+from .logs import logs
+from .help import help
+from .config import config
+from .notebooks import notebooks
+from .transforms import transforms
