--- viewser/formatting/__init__.py
@@ -1,1 +1,0 @@
-from .documentation_formatter import DocumentationFormatter
