--- viewser/formatting/database_table_formatter.py
@@ -1,1 +1,0 @@
-
