+++ viewser/hirearchical_dict.py
@@ -1,0 +1,9 @@
+
+class HirearchicalDict(dict):
+    def __setitem__(self, key, value):
+        try:
+            self[key]
+        except KeyError:
+            super().__setitem__(key, value)
+        else:
+            pass
