--- viewser/models.py
@@ -1,2 +1,0 @@
-
-from views_schema import *
