--- viewser/schema.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-class IRemotePaths(Enum):
-    querysets = 1
-    documentation = 2
