+++ viewser/tui/animations.py
